+++ src/ScottPlot5/ScottPlot5 Demos/ScottPlot5 WinForms Demo/Program.cs
@@ -12,9 +12,9 @@ static class Program
         // use this to quickly launch a test Form while developing
         if (true && Environment.MachineName == "DESKTOP-L7MMAB7")
         {
-            Demos.MouseTracker window = new() { StartPosition = FormStartPosition.CenterScreen };
+            Demos.DraggableCallout window = new() { StartPosition = FormStartPosition.CenterScreen };
             Application.Run(window);
-            Application.Run(new MainMenuForm());
+            //Application.Run(new MainMenuForm());
         }
         else
         {