+++ CHANGELOG.md
@@ -13,6 +13,7 @@ _Not yet on NuGet..._
 * WinUI: Added automatic display scaling detection and correction (#3642) @PZidlik
 * Bar: Added a `CenterLabel` flag to cause value labels to be displayed centered within a bar (#3391) @tibormarchynzoom
 * FormsPlot: Allow plots to persist through `Show()` and `Close()` events (#3643, #3589) @CodeBehemoth @bwedding @Kruno313
+* Callout: New plot type that displays text with an arrow that points to a location on the plot (#3650, #3654) @NicolasLairNET
 
 ## ScottPlot 4.1.74
 * Security: Use System.Drawing.Common version 4.7.2 to address [CVE-2021-26701](https://github.com/advisories/GHSA-ghhp-997w-qr28)